--- a/tests/phpunit/testdata/plugins/test-plugin-localhost-without-errors/load.php
+++ b/tests/phpunit/testdata/plugins/test-plugin-localhost-without-errors/load.php
@@ -18,5 +18,3 @@
  /*
   * This file contains https://localhost/test-plugin-localhost-with-errors url.
   */
-
-$file = 'https://localhost/test-plugin-localhost-with-errors';