### Description of the Change

In #267, @leoloso reported that the error "Do not use Localhost in your code" also gets triggered when the string "localhost" is used within a comment. This PR updates the corresponding regex to ignore the string "localhost" within comments.

Closes #267

### How to test the Change

1. Download and activate this [helper-plugin.zip](https://github.com/10up/plugin-check/files/12708275/helper-plugin.zip)
2. Go to `/wp-admin/tools.php?page=plugin-check` and check the helper plugin before applying this fix.
3. Verify that the error "Do not use Localhost/127.0.0.1 in your code." appears.
4. Apply this PR and check the helper plugin again.
5. Verify that the error "Do not use Localhost/127.0.0.1 in your code." no longer appears.

### Changelog Entry

> Fixed - “localhost” in comments no longer raises error.

### Credits

Props @nielslange

### Checklist:
- [x] I agree to follow this project's [**Code of Conduct**](https://github.com/10up/.github/blob/trunk/CODE_OF_CONDUCT.md).
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my change.
- [x] All new and existing tests pass.