@ernilambar 2023-12-28
I used `time` command to find the time taken for completion of the command. But it seems `time` gives different output for same command run consecutively. So, I ran commands 3 times and calculate the average time. This does not seem to be reliable test but I did not find much significant drops in performance when trying with three big WordPress plugins.

Before PR:

* WooCommerce - 7.54s
* Yoast SEO - 2.67s
* EDD - 4.99s

After PR:

* WooCommerce - 6.45s
* Yoast SEO - 3.12s
* EDD - 3.76s