@ernilambar 2023-12-28
While checking this PR, I found another issue. https://github.com/WordPress/plugin-check/issues/362