@davidperezgar 2024-10-20
Yes, I'm agree with @ernilambar