@ernilambar 2024-10-24
@nielslange Thanks for the PR and contribution. Closing this now as sniff approach looked more appropriate now.

Closing in favor of https://github.com/WordPress/plugin-check/pull/743